refactor: move useStreami18n hook to hooks folder
