+++ package/src/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './useAppStateListener';
+export * from './useStreami18n';
